backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.0
openai>=1.0.0
